Turret/Gun System (#15)
Added a flak turret with AI, press 'A' to add it to the scene, and 'T' to access the Turrets Menu.

Turrets are placed with the same position and rotation as set GameObjects with the TurretSlot script. TurretSlot also has a RotationLimit component which controls how much the attached turret can rotate on the x (vertical, gun) and y (horizontal, turret) axes.

Turret rotation limits can range from -360 to 360, but the min and max range must be limited within 360 degrees of each other. They must also pass through 0, which is the default starting rotation of every turret. The turret rotates from min to max, and the value chosen for min and max determines if it rotates clockwise or anti-clockwise. Place the turret to test it and find the values you want.

All Damageable objects contain a UnitProperties struct, which help identify which Team (Player or Enemy) the object belongs to, and what UnitType enum flag it is (Ground, Flying, Melee, Ranged, others examples could be Normal, Elite, Shielded, Artillery, Drone, Turret, etc...)

AI turrets contain a TargetFinder, which uses a trigger Sphere Collider to find targets within range of it. You can also specify a soft preference for which targets are valid. Soft preferences are matches of Team values and if any of the UnitType enum flag are a match. Hard criteria are also matches of Team values, but all the UnitType enum flags must be a match from the preference. If the UnitType is left as None, this will always be implemented the same as if every flag was selected (Everything).

The AI Turret Controller then handles the rotation and targeting of the turret. This AI can have multiple targeting criteria, and is ordered by priority. For example, in the picture below, the AI will prioritise targeting targets that are Flying over Ground, but will still target Ground targets if there are no Flying enemies in range and in rotation range. If a target does not meet any criteria, it will be assigned the lowest priority.

Projectiles now also have an Effective Against UnitType value. This determines whether its collisions would result in full damage, and if splash damage will occur. In the image below, the Projectile is only effective against flying types (also soft preference), and if it hit an object that was not of flying type, it would deal a smaller amount of damage, and splash damage would not proc if it was enabled. The weakened damage percentage can be set in the Damage SO.