AiMovement refactoring with GroupStat variables, adjustment of max speed handling
Aimovementcontroller now looks at the speed + the delta of the acceleration to see whether it exceeds the speed, otherwise, objects aren't able to receive force when there is no friction. This is because the lack of friction will never allow the speed to drop, meaning that even if you try to apply an opposite force, the controller wouldn't allow it because the velocity is still too high. Now you can brake with opposite force