Added an Enemy Director, ScriptableObject event system, HP and EXP, and Upgrades that trigger on level up.
Enemy Director
The Enemy Director spawns in intervals, and checks for Spawn Locations (empty objects with a trigger collider with its own layer mask) in a distance around the player, and removes those that are too close to the player. It then chooses a random Enemy to spawn at one of those Spawn Location.

ScriptableObject Event System
Implementation from: https://www.youtube.com/watch?v=iXNwWpG7EhM

Utilises a GameEvent ScriptableObject, a MonoBehaviour Listener, and a UnityEvent Response. Each of these are generics of a type. To create an event and response of a type, a script of each type has to be created. See the GameEvents folder under Scripts. To create a GameEvent, create the ScriptableObject of the type required, reference it in another script, and call Raise of that asset, passing in the required parameter. To listen for this event, attach a Listener of the type required to a gameobject, and assign the UnityEvent response on it.